renamed Form to SignUpForm
