+++ quick-tips/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Form from './components/Form';
+import Form from './components/SignUpForm';
 
 function App() {
   return (
