+++ quick-tips/src/components/SignUpForm.css
@@ -40,7 +40,7 @@
     display: flex;
     flex-direction: column;
     width: 60%;
-    height: 65vh;
+    min-height: 55vh;
     align-self: center;
     box-shadow: 2px 2px 8px lightgray;
 }
@@ -83,3 +83,4 @@
         width: 75vh;
     }
 }
+@media(min-width:)
